refactor: move getUnsupportedTokensFromPermitInfo to utils fn
